fix: fix crash issue
