+++ PEAK-AIO/Main.cs
@@ -19,7 +19,6 @@ public class PeakMod : BaseUnityPlugin
 {
     // Menu
     private bool styleApplied = false;
-    private bool fontsLoaded = false;
     private bool showMenu = false;
     private int selectedTab = 1;
     private static readonly FieldInfo cursorVisibleField = typeof(DearImGuiInjection.DearImGuiInjection)
@@ -106,66 +105,6 @@ public class PeakMod : BaseUnityPlugin
         style.ItemSpacing = new System.Numerics.Vector2(2, 4);
     }
 
-    private unsafe void LoadCJKFonts()
-    {
-        var io = ImGui.GetIO();
-        var fonts = io.Fonts;
-
-        fonts.Clear();
-        fonts.AddFontDefault();
-
-        float fontSize = 13.0f;
-
-        ImFontConfigPtr mergeConfig = new ImFontConfigPtr(ImGuiNative.ImFontConfig_ImFontConfig());
-        mergeConfig.MergeMode = true;
-        mergeConfig.PixelSnapH = true;
-
-        string msyhPath = @"C:\Windows\Fonts\msyh.ttc";
-        if (System.IO.File.Exists(msyhPath))
-        {
-            fonts.AddFontFromFileTTF(msyhPath, fontSize, mergeConfig, fonts.GetGlyphRangesChineseFull());
-            fonts.AddFontFromFileTTF(msyhPath, fontSize, mergeConfig, fonts.GetGlyphRangesJapanese());
-        }
-
-        string malgunPath = @"C:\Windows\Fonts\malgun.ttf";
-        if (System.IO.File.Exists(malgunPath))
-        {
-            fonts.AddFontFromFileTTF(malgunPath, fontSize, mergeConfig, fonts.GetGlyphRangesKorean());
-        }
-
-        fonts.Build();
-        mergeConfig.Destroy();
-
-        var assembly = typeof(DearImGuiInjection.DearImGuiInjection).Assembly;
-        string[] backendTypeNames = {
-            "DearImGuiInjection.Backends.ImGuiDX12Impl",
-            "DearImGuiInjection.Backends.ImGuiDX11Impl"
-        };
-
-        foreach (var typeName in backendTypeNames)
-        {
-            try
-            {
-                var implType = assembly.GetType(typeName);
-                if (implType == null) continue;
-
-                var deviceField = implType.GetField("_device", BindingFlags.NonPublic | BindingFlags.Static);
-                if (deviceField == null || deviceField.GetValue(null) == null) continue;
-
-                var method = implType.GetMethod("CreateFontsTexture", BindingFlags.Public | BindingFlags.Static);
-                if (method == null) continue;
-
-                method.Invoke(null, null);
-                Logger.LogInfo($"[PEAK AIO] Font texture recreated via {typeName}");
-                break;
-            }
-            catch (Exception ex)
-            {
-                Logger.LogWarning($"[PEAK AIO] {typeName} font texture failed: {ex.Message}");
-            }
-        }
-    }
-
     private void Awake()
     {
         Logger.LogInfo("Mod Initialized");
@@ -180,9 +119,38 @@ public class PeakMod : BaseUnityPlugin
         ConfigManager.Init(Config, Logger);
         DearImGuiInjection.DearImGuiInjection.Render += MyUI;
 
-        // Initialize Harmony
         var harmony = new Harmony("com.onigremlin.peakaio");
         harmony.PatchAll();
+
+        var diAssembly = typeof(DearImGuiInjection.DearImGuiInjection).Assembly;
+        var prefixMethod = new HarmonyMethod(typeof(CJKFontPatch).GetMethod("Prefix",
+            BindingFlags.Public | BindingFlags.Static));
+
+        string[] backendTypeNames = {
+            "DearImGuiInjection.Backends.ImGuiDX12Impl",
+            "DearImGuiInjection.Backends.ImGuiDX11Impl"
+        };
+
+        foreach (var typeName in backendTypeNames)
+        {
+            try
+            {
+                var implType = diAssembly.GetType(typeName);
+                if (implType == null) continue;
+
+                var newFrameMethod = implType.GetMethod("NewFrame",
+                    BindingFlags.Public | BindingFlags.NonPublic | BindingFlags.Static);
+                if (newFrameMethod == null) continue;
+
+                harmony.Patch(newFrameMethod, prefix: prefixMethod);
+                Logger.LogInfo($"[PEAK AIO] Patched {typeName}.NewFrame for CJK fonts.");
+            }
+            catch (Exception ex)
+            {
+                Logger.LogWarning($"[PEAK AIO] Failed to patch {typeName}.NewFrame: {ex.Message}");
+            }
+        }
+
         Logger.LogInfo("Harmony patches applied.");
     }
 
@@ -295,21 +263,6 @@ public class PeakMod : BaseUnityPlugin
     {
         try
         {
-            if (!fontsLoaded)
-            {
-                fontsLoaded = true;
-                try
-                {
-                    LoadCJKFonts();
-                    Logger.LogInfo("[PEAK AIO] CJK font atlas rebuilt.");
-                }
-                catch (Exception fontEx)
-                {
-                    Logger.LogWarning("[PEAK AIO] CJK font loading failed, using default font: " + fontEx.Message);
-                }
-                return;
-            }
-
             if (!showMenu)
                 return;
